Added tests for moving the hero to a free cell, against a wall and to the same point as the Sword
