+++ src/logic/Actor.java
@@ -1,4 +1,4 @@
-package cli;
+package logic;
 
 public class Actor {
