+++ src/logic/Dragon.java
@@ -1,4 +1,4 @@
-package cli;
+package logic;
 
 import java.util.Random;
