+++ src/logic/Game.java
@@ -1,4 +1,4 @@
-package cli;
+package logic;
 
 import java.util.Random;
 
@@ -11,7 +11,70 @@ public class Game {
 	
 	//Constructor
 	public Game(char type) {
-		myMaze = new Maze();
+		char [][] grid;
+		grid = new char[10][10];
+
+		// Linha 0
+
+		for (int i = 0; i < 10; i++)
+			grid[0][i] = 'X';
+
+		// Linha 9
+		for (int i = 0; i < 10; i++)
+			grid[9][i] = 'X';
+
+		// Coluna 1
+
+		for (int i = 0; i < 10; i++) {
+			grid[i][0] = 'X';
+		}
+
+		// Coluna 9
+
+		for (int i = 0; i < 10; i++) {
+			grid[i][9] = 'X';
+		}
+
+		grid[2][2] = 'X';
+		grid[2][3] = 'X';
+		grid[2][5] = 'X';
+		grid[2][7] = 'X';
+
+		grid[3][2] = 'X';
+		grid[3][3] = 'X';
+		grid[3][5] = 'X';
+		grid[3][7] = 'X';
+
+		grid[4][2] = 'X';
+		grid[4][3] = 'X';
+		grid[4][5] = 'X';
+		grid[4][7] = 'X';
+
+		grid[6][2] = 'X';
+		grid[6][3] = 'X';
+		grid[6][5] = 'X';
+		grid[6][7] = 'X';
+
+		grid[7][2] = 'X';
+		grid[7][3] = 'X';
+		grid[7][5] = 'X';
+		grid[7][7] = 'X';
+
+		grid[5][7] = 'X';
+
+		grid[8][2] = 'X';
+		grid[8][3] = 'X';
+
+		for (int i = 0; i < 10; i++) {
+			for (int j = 0; j < 10; j++) {
+				if (grid[i][j] != 'X')
+					grid[i][j] = ' ';
+			}
+		}
+		grid[5][9] = 'S';
+	
+		
+		myMaze = new Maze(grid);
 		myHero = new Hero(1,1,'H');
 		mySword = new Sword(8,1,'E');
 		myDragon = new Dragon(3,1,'D');
@@ -20,6 +83,7 @@ public class Game {
 	
 	//Starts Game
 	public void play(char m){
+		
 		
 		boolean enableSleep;
 		myMaze.removeActor(myHero);
